core 폴더 추가
