+++ src/core/index.ts
@@ -1,0 +1,3 @@
+export * from './bezier-curve';
+export * from './lerp';
+export * from './observer';
