+++ src/lib/index.ts
@@ -1,8 +1,5 @@
 export * from './canvas-utils';
 export * from './config';
 export * from './controller';
-export * from './lerp';
-export * from './observer';
 export * from './presets';
 export * from './renderer';
-export * from './types';
